+++ stdlib/public/CMakeLists.txt
@@ -79,6 +79,8 @@ if(SWIFT_BUILD_STDLIB OR SWIFT_BUILD_REMOTE_MIRROR)
     "${SWIFT_SOURCE_DIR}/lib/Demangling/Remangler.cpp"
     "${SWIFT_SOURCE_DIR}/lib/Demangling/NodeDumper.cpp"
     "${SWIFT_SOURCE_DIR}/lib/Demangling/Errors.cpp")
+  set(swiftDemanglingCRSources
+    "${SWIFT_SOURCE_DIR}/lib/Demangling/CrashReporter.cpp")
 
   set(swift_demangling_cflags)
 
@@ -109,11 +111,10 @@ if(SWIFT_BUILD_STDLIB OR SWIFT_BUILD_REMOTE_MIRROR)
       ${swift_demangling_cflags}
     INSTALL_IN_COMPONENT never_install)
 
-  add_swift_target_library(swiftDemanglingForCore OBJECT_LIBRARY
-    ${swiftDemanglingSources}
+  add_swift_target_library(swiftDemanglingCR OBJECT_LIBRARY
+    ${swiftDemanglingCRSources}
     C_COMPILE_FLAGS
       -DswiftCore_EXPORTS
-      -DSWIFT_DEMANGLE_USE_RUNTIME_ERRORS=1
       ${swift_demangling_cflags}
     INSTALL_IN_COMPONENT never_install)
 endif()
